pyproject.toml
[project]
name = "pyside-yolov10"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26.4",
    "onnxruntime-gpu>=1.19.2",
    "opencv-python>=4.10.0.84",
    "pyside6>=6.8.0.2",
    "torch==2.5.1+cu121",
    "torchaudio==2.5.1+cu121",
    "torchvision==0.20.1+cu121",
    "typing-extensions>=4.12.2",
]

[tool.uv]
index-strategy = "unsafe-best-match"
index-url = "https://pypi.org/simple"
extra-index-url = [
    "https://pypi.org/simple",
    "https://download.pytorch.org/whl/cu121",
]
dev-dependencies = [
    "dill>=0.3.9",
    "ipykernel>=6.29.5",
    "nuitka>=2.4.11",
    "pre-commit>=4.0.1",
    "ruff>=0.7.1",
    "rust-just>=1.36.0",
    "tensorboard>=2.18.0",
    "ultralytics>=8.3.25",
    "wandb>=0.18.5",
]
